requirements.txt
python-pptx==0.6.23
openai==1.12.0
langchain==0.1.10
langchain-openai==0.0.5
python-dotenv==1.0.0

